+++ src/Haskellorls/Formatter/Escape.hs
@@ -1,12 +1,12 @@
-module Haskellorls.Formatter.Escape (escapeFormatter) where
+module Haskellorls.Formatter.Escape (escape) where
 
 import Data.Char
 import qualified Data.Text as T
 import qualified Haskellorls.Config as Config
 import qualified Haskellorls.Config.Quote as Quote
 
-replaceControlCharsToQuestion :: T.Text -> T.Text
-replaceControlCharsToQuestion = T.map (\c -> if isPrint c then c else '?')
+replaceControlCharsByQuestion :: T.Text -> T.Text
+replaceControlCharsByQuestion = T.map (\c -> if isPrint c then c else '?')
 
 escapeCharsForStdout :: T.Text -> T.Text
 escapeCharsForStdout = T.concatMap $ \case
@@ -14,26 +14,33 @@ escapeCharsForStdout = T.concatMap $ \case
   '\t' -> "'$'\\t''"
   c -> T.singleton c
 
--- | NOTE: There may be some missing targets.
-escapeCharsForStdoutByCStyle :: T.Text -> T.Text
-escapeCharsForStdoutByCStyle = T.concatMap $ \case
+-- | The @'escapeAsCLiteral' b t@ function substitutes characters in 'Text'
+-- with C string literal expression. @b@ indicates whether or not the
+-- substituted string will be surrounded by double quotes. If it is 'True' this
+-- function also substitutes @"@ with an appropriate expression in addition to
+-- standard targets, otherwise also substitutes @ @ (a space) with an
+-- appropriate one too.
+--
+-- NOTE: Should we substitute other non-printable characters? And can these
+-- characters be contained in a valid filepath?
+escapeAsCLiteral :: Bool -> T.Text -> T.Text
+escapeAsCLiteral willQuote = T.concatMap $ \case
+  '"' | willQuote -> "\""
+  ' ' | not willQuote -> "\\ "
   '\t' -> "\\t"
   '\r' -> "\\r"
   '\n' -> "\\n"
-  ' ' -> "\\ "
-  '|' -> "\\|"
   '\\' -> "\\\\"
   c -> T.singleton c
 
-escapeFormatter :: Config.Config -> T.Text -> T.Text
-escapeFormatter config = case Config.quotingStyle config of
-  Quote.Literal -> replaceControlCharsToQuestion
-  Quote.Shell -> replaceControlCharsToQuestion
-  Quote.ShellAlways -> replaceControlCharsToQuestion
+escape :: Config.Config -> T.Text -> T.Text
+escape config = case Config.quotingStyle config of
+  Quote.Literal
+    | Config.showControlChars config -> id
+    | otherwise -> replaceControlCharsByQuestion
+  Quote.Shell -> replaceControlCharsByQuestion
+  Quote.ShellAlways -> replaceControlCharsByQuestion
   Quote.ShellEscape -> escapeCharsForStdout
   Quote.ShellEscapeAlways -> escapeCharsForStdout
-  Quote.C -> escapeCharsForStdoutByCStyle
-  Quote.Escape -> escapeCharsForStdoutByCStyle
-  Quote.NoStyle
-    | Config.toTTY config -> escapeCharsForStdout
-    | otherwise -> id
+  Quote.C -> escapeAsCLiteral True
+  Quote.Escape -> escapeAsCLiteral False